% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CelltypeNetwork-class}
\alias{CelltypeNetwork-class}
\title{CelltypeNetwork: virtual base class for cell-type networks}
\description{
CelltypeNetwork: virtual base class for cell-type networks
}
\examples{
showClass("CelltypeNetwork")
}
