% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TemporalNetwork-class}
\alias{TemporalNetwork-class}
\alias{show,TemporalNetwork-method}
\title{TemporalNetwork: temporal network evolution result}
\usage{
\S4method{show}{TemporalNetwork}(object)
}
\arguments{
\item{object}{A \code{TemporalNetwork} object.}
}
\value{
A \code{TemporalNetwork} object.
}
\description{
Stores results from \code{\link{run_temporal_network}},
tracking edge dynamics across time points.
}
\section{Slots}{

\describe{
\item{\code{networks}}{List of \code{CelltypeNetworkResult} objects,
one per time point.}

\item{\code{edge_dynamics}}{\code{DataFrame}. Edge-level temporal
trajectory: node_i, node_j, status at each time point.}

\item{\code{timepoints}}{Character vector. Ordered time point labels.}
}}

\examples{
showClass("TemporalNetwork")
}
