% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{exposureData}
\alias{exposureData}
\alias{exposureInfo}
\alias{sampleMap}
\alias{exposureNames}
\alias{exposureData<-}
\alias{exposureInfo<-}
\alias{exposureData,SingleCellExposomeExperiment-method}
\alias{exposureInfo,SingleCellExposomeExperiment-method}
\alias{sampleMap,SingleCellExposomeExperiment-method}
\alias{exposureNames,SingleCellExposomeExperiment-method}
\alias{exposureData<-,SingleCellExposomeExperiment-method}
\alias{exposureInfo<-,SingleCellExposomeExperiment-method}
\title{Access exposure data matrix}
\usage{
exposureData(x)

exposureInfo(x)

sampleMap(x)

exposureNames(x)

exposureData(x) <- value

exposureInfo(x) <- value

\S4method{exposureData}{SingleCellExposomeExperiment}(x)

\S4method{exposureInfo}{SingleCellExposomeExperiment}(x)

\S4method{sampleMap}{SingleCellExposomeExperiment}(x)

\S4method{exposureNames}{SingleCellExposomeExperiment}(x)

\S4method{exposureData}{SingleCellExposomeExperiment}(x) <- value

\S4method{exposureInfo}{SingleCellExposomeExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{SingleCellExposomeExperiment}}.}

\item{value}{A \code{DataFrame}.}
}
\value{
Numeric matrix (samples x exposures).

A \code{DataFrame}.

A \code{DataFrame}.

Character vector.
}
\description{
Access exposure data matrix

Access exposure metadata

Access sample-cell mapping

Get exposure variable names

Replace exposure data matrix

Replace exposure metadata
}
\examples{
library(SingleCellExperiment)
counts <- matrix(1L, nrow = 5, ncol = 12,
    dimnames = list(paste0("G", 1:5), paste0("c", 1:12)))
sce <- SingleCellExperiment(assays = list(counts = counts),
    colData = S4Vectors::DataFrame(
        cell_id = paste0("c", 1:12),
        donor_id = rep(paste0("D", 1:4), each = 3)))
exp_mat <- matrix(c(12, 25, 8, 30, 45, 32, 67, 51), nrow = 4,
    dimnames = list(paste0("D", 1:4), c("PM2.5", "age")))
scee <- build_scee(sce, exp_mat, sample_col = "donor_id")

exposureNames(scee)
exposureData(scee)
exposureInfo(scee)
head(sampleMap(scee))

## Replace the exposure matrix, here rescaling PM2.5
exp_new <- exposureData(scee)
exp_new[, "PM2.5"] <- exp_new[, "PM2.5"] / 10
exposureData(scee) <- exp_new
exposureData(scee)[, "PM2.5"]
}
