% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{exposureData}
\alias{exposureData}
\alias{exposureInfo}
\alias{sampleMap}
\alias{exposureNames}
\alias{exposureData<-}
\alias{exposureInfo<-}
\alias{exposureData,SingleCellExposomeExperiment-method}
\alias{exposureInfo,SingleCellExposomeExperiment-method}
\alias{sampleMap,SingleCellExposomeExperiment-method}
\alias{exposureNames,SingleCellExposomeExperiment-method}
\alias{exposureData<-,SingleCellExposomeExperiment-method}
\alias{exposureInfo<-,SingleCellExposomeExperiment-method}
\title{Access exposure data matrix}
\usage{
exposureData(x)

exposureInfo(x)

sampleMap(x)

exposureNames(x)

exposureData(x) <- value

exposureInfo(x) <- value

\S4method{exposureData}{SingleCellExposomeExperiment}(x)

\S4method{exposureInfo}{SingleCellExposomeExperiment}(x)

\S4method{sampleMap}{SingleCellExposomeExperiment}(x)

\S4method{exposureNames}{SingleCellExposomeExperiment}(x)

\S4method{exposureData}{SingleCellExposomeExperiment}(x) <- value

\S4method{exposureInfo}{SingleCellExposomeExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{SingleCellExposomeExperiment}}.}

\item{value}{A \code{DataFrame}.}
}
\value{
Numeric matrix (samples x exposures).

A \code{DataFrame}.

A \code{DataFrame}.

Character vector.
}
\description{
Access exposure data matrix

Access exposure metadata

Access sample-cell mapping

Get exposure variable names

Replace exposure data matrix

Replace exposure metadata
}
\examples{
showClass("SingleCellExposomeExperiment")
}
